堂食订单，外卖订单后端开发，前端堂食订单页面初始化

+++ sz-boot-parent/sz-service/sz-service-admin/src/main/java/com/sz/admin/restaurant/pojo/dto/DineInOrdersListDTO.java
@@ -16,9 +16,8 @@ import com.sz.core.common.entity.PageQuery;
 public class DineInOrdersListDTO extends PageQuery {
 
     @Schema(description =  "")
-    private Integer orderId;
+    private Long orderId;
 
     @Schema(description =  "")
-    private Integer tableId;
-
+    private Long tableId;
 }
